# openapi.yaml
openapi: 3.1.0
info:
  title: Trustify
  description: Software Supply-Chain Security API
  license:
    name: Apache License, Version 2.0
    identifier: Apache-2.0
  version: 0.1.1
paths:
...
  /api/v1/ai/completions:
    post:
      tags:
      - ai
      operationId: completions
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatState'
        required: true
      responses:
        '200':
          description: The resulting completion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatState'
        '400':
          description: The request was invalid
        '404':
          description: The AI service is not enabled
  /api/v1/ai/conversations:
    get:
      tags:
      - ai
      operationId: listConversations
      parameters:
      - name: q
        in: query
        required: false
        schema:
          type: string
      - name: sort
        in: query
        required: false
        schema:
          type: string
      - name: offset
        in: query
        description: |-
          The first item to return, skipping all that come before it.

          NOTE: The order of items is defined by the API being called.
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      - name: limit
        in: query
        description: |-
          The maximum number of entries to return.

          Zero means: no limit
        required: false
        schema:
          type: integer
          format: int64
          minimum: 0
      responses:
        '200':
          description: The resulting list of conversation summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResults_ConversationSummary'
        '404':
          description: The AI service is not enabled
    post:
      tags:
      - ai
      operationId: createConversation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatState'
        required: true
      responses:
        '200':
          description: The resulting conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          description: The request was invalid
        '404':
          description: The AI service is not enabled
  /api/v1/ai/conversations/{id}:
    get:
      tags:
      - ai
      operationId: getConversation
      parameters:
      - name: id
        in: path
        description: Opaque ID of the conversation
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: The resulting conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          description: The request was invalid
        '404':
          description: The AI service is not enabled or the conversation was not found
    put:
      tags:
      - ai
      operationId: updateConversation
      parameters:
      - name: id
        in: path
        description: Opaque ID of the conversation
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Conversation'
        required: true
      responses:
        '200':
          description: The resulting conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          description: The request was invalid
        '404':
          description: The AI service is not enabled or the conversation was not found
    delete:
      tags:
      - ai
      operationId: deleteConversation
      parameters:
      - name: id
        in: path
        description: Opaque ID of the conversation
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: The resulting conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          description: The request was invalid
        '404':
          description: The AI service is not enabled or the conversation was not found
  /api/v1/ai/flags:
    get:
      tags:
      - ai
      operationId: aiFlags
      responses:
        '200':
          description: The resulting Flags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiFlags'
        '404':
          description: The AI service is not enabled
  /api/v1/ai/tools:
    get:
      tags:
      - ai
      operationId: aiTools
      responses:
        '200':
          description: The resulting list of tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AiTool'
        '404':
          description: The AI service is not enabled
  /api/v1/ai/tools/{name}:
    post:
      tags:
      - ai
      operationId: aiToolCall
      parameters:
      - name: name
        in: path
        description: Name of the tool to call
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema: {}
        required: true
      responses:
        '200':
          description: The result of the tool call
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: The tool request was invalid
        '404':
          description: The tool was not found
...
components:
  schemas:
...
    AiFlags:
      type: object
      required:
      - completions
      properties:
        completions:
          type: boolean
    AiTool:
      type: object
      required:
      - name
      - description
      - parameters
      properties:
        description:
          type: string
        name:
          type: string
        parameters: {}
...
    ChatMessage:
      type: object
      required:
      - message_type
      - content
      properties:
        content:
          type: string
        internal_state:
          type:
          - string
          - 'null'
        message_type:
          $ref: '#/components/schemas/MessageType'
    ChatState:
      type: object
      required:
      - messages
      properties:
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
...
    Conversation:
      type: object
      required:
      - id
      - state
      - updated_at
      - seq
      properties:
        id:
          type: string
          format: uuid
        seq:
          type: integer
          format: int32
        state:
          $ref: '#/components/schemas/ChatState'
        updated_at:
          type: string
          format: date-time
...
    MessageType:
      type: string
      enum:
      - human
      - system
      - ai
      - tool
...
    PaginatedResults_ConversationSummary:
      type: object
      required:
      - items
      - total
      properties:
        items:
          type: array
          items:
            type: object
            required:
            - id
            - updated_at
            - summary
            properties:
              id:
                type: string
                format: uuid
              summary:
                type: string
              updated_at:
                type: string
                format: date-time
        total:
          type: integer
          format: int64
          minimum: 0
...